@@ -14,9 +14,9 @@
     "dist": "webpack --progress --colors --config build/dist.config.js",
     "watch-dist": "webpack --watch --progress --colors --config build/dist.config.js",
     "dist.min": "webpack --progress --colors -p --config build/dist.min.config.js",
-    "build": "NODE_ENV=production npm run lib && npm run dist && npm run dist.min && npm run build-style && npm run build-style-dist && npm run build-style-dist-min",
-    "dev": "NODE_ENV=production PORT=9090 webpack-dev-server --progress --colors --config  build/dev.config.js",
-    "hot": "NODE_ENV=production HOT=true PORT=9090 webpack-dev-server --progress --colors --config build/dev.config.js"
+    "build": "SET NODE_ENV=production && npm run lib && npm run dist && npm run dist.min && npm run build-style && npm run build-style-dist && npm run build-style-dist-min",
+    "dev": "SET NODE_ENV=production && SET PORT=9090 && webpack-dev-server --progress --colors --config  build/dev.config.js",
+    "hot": "SET NODE_ENV=production && SET HOT=true && SET PORT=9090 && webpack-dev-server --progress --colors --config build/dev.config.js"
   },
   "repository": {
     "type": "git",
@@ -40,7 +40,7 @@
     "babel-loader": "^6.1.0",
     "babel-preset-es2015": "^6.1.2",
     "babel-preset-react": "^6.1.2",
-    "css-loader": "^0.9.0",
+    "css-loader": "^0.23.1",
     "exports-loader": "^0.6.2",
     "extract-text-webpack-plugin": "^0.8.2",
     "faker": "^3.0.1",
@@ -53,12 +53,14 @@
     "node-libs-browser": "^0.5.2",
     "parse-keys": "^1.3.0",
     "raw-loader": "^0.5.1",
-    "react-dom": "^0.14.2",
+    "react": "^15.1.0",
+    "react-dom": "^15.1.0",
     "react-hot-loader": "^1.2.7",
     "should": "^7.1.1",
     "sorty": "^1.2.0",
     "style-loader": "^0.13.0",
-    "stylus-loader": "^1.4.0",
+    "stylus": "^0.54.5",
+    "stylus-loader": "^2.1.1",
     "webpack": "^1.9.12",
     "webpack-dev-server": "^1.9.0"
   },
@@ -72,11 +74,9 @@
     "node-uuid": "^1.4.3",
     "normalize.css": "^3.0.3",
     "object-assign": "^4.0.1",
-    "react": "^0.14.0",
     "react-class": "^1.2.4",
-    "react-dom": "^0.14.0",
-    "react-load-mask": "^1.0.1",
-    "react-menus": "^2.0.4",
+    "react-load-mask": "^2.0.1",
+    "react-menus": "^2.0.5",
     "react-simple-toolbar": "^1.0.5",
     "react-style-normalizer": "^1.2.8",
     "react-virtual-scroller": "^2.1.2-lic.4",
@@ -86,8 +86,8 @@
     "whatwg-fetch": "^0.10.0"
   },
   "peerDependencies": {
-    "react": ">=0.14.0",
-    "react-dom": ">=0.14.0"
+    "react": ">=0.14.0 || >=15.0.0",
+    "react-dom": ">=0.14.0 || >=15.0.0"
   },
   "author": "ZippyUI <contact@zippyui.com>",
   "contributors": [],
