@@ -7,7 +7,7 @@ import { findDOMNode } from 'react-dom'
 import React from 'react'
 
 var assign   = require('object-assign')
-var LoadMask = require('react-load-mask')
+import LoadMask from 'react-load-mask'
 var Region   = require('region')
 
 var PaginationToolbar = React.createFactory(require('./PaginationToolbar'))
@@ -1102,7 +1102,7 @@ module.exports = React.createClass({
 
     onColumnResizeDragStart: function(config){
 
-        var domNode = this.getDOMNode()
+        var domNode = findDOMNode(this)
         var region  = Region.from(domNode)
 
         this.resizeProxyLeft = config.resizeProxyLeft - region.left
