@@ -3,6 +3,7 @@
 var React = require('react');
 var assign = require('object-assign');
 var ReactMenu = require('react-menus');
+var findDOMNode = require('react-dom').findDOMNode;
 
 function stopPropagation(event) {
     event.stopPropagation();
@@ -175,7 +176,7 @@ module.exports = {
     },
 
     onFilterMenuMount: function onFilterMenuMount(menu) {
-        var dom = menu.getDOMNode();
+        var dom = findDOMNode(menu);
 
         if (dom) {
             var input = dom.querySelector('input');
