@@ -315,7 +315,7 @@ module.exports = React.createClass({
 
     showMenu: function showMenu(column, event) {
 
-        var menuItem = (function (column) {
+        var menuItem = function (column) {
             var visibility = this.props.columnVisibility;
 
             var visible = column.visible;
@@ -334,7 +334,7 @@ module.exports = React.createClass({
                 label: column.title,
                 fn: this.toggleColumn.bind(this, column)
             };
-        }).bind(this);
+        }.bind(this);
 
         function menu(eventTarget, props) {
 
@@ -451,6 +451,9 @@ module.exports = React.createClass({
             return;
         }
 
+        if (!column.draggable) {
+            return;
+        }
         setupColumnDrag(this, this.props, column, event);
     },
 
