@@ -2,6 +2,7 @@
 
 var Region = require('region');
 var DragHelper = require('drag-helper');
+var findDOMNode = require('react-dom').findDOMNode;
 
 function range(start, end) {
     var res = [];
@@ -29,7 +30,7 @@ module.exports = function (header, props, column, event) {
 
     event.preventDefault();
 
-    var headerNode = header.getDOMNode();
+    var headerNode = findDOMNode(header);
     var headerRegion = Region.from(headerNode);
     var dragColumn = column;
     var dragColumnIndex;
