@@ -251,11 +251,11 @@ module.exports = React.createClass({
 		var page = this.props.page;
 		this.setState({
 			inputFocused: true
-		}, (function () {
+		}, function () {
 
 			var domNode = (0, _reactDom.findDOMNode)(this.refs.input);
 			domNode.value = page;
-		}).bind(this));
+		}.bind(this));
 	},
 
 	onPageSizeChange: function onPageSizeChange(event) {
