@@ -6,12 +6,16 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _reactLoadMask = require('react-load-mask');
+
+var _reactLoadMask2 = _interopRequireDefault(_reactLoadMask);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 require('es6-promise').polyfill();
 
 var assign = require('object-assign');
-var LoadMask = require('react-load-mask');
+
 var Region = require('region');
 
 var PaginationToolbar = _react2.default.createFactory(require('./PaginationToolbar'));
@@ -232,8 +236,8 @@ module.exports = _react2.default.createClass({
 
             // console.log(scrollTop, sign);
         } else {
-                state.scrollTop = scrollTop;
-            }
+            state.scrollTop = scrollTop;
+        }
 
         this.setState(state);
     },
@@ -324,11 +328,11 @@ module.exports = _react2.default.createClass({
             });
         }
 
-        setTimeout((function () {
+        setTimeout(function () {
             //since menu is hidden on click on window,
             //show it in a timeout, after the click event has reached the window
             this.setState(state);
-        }).bind(this), 0);
+        }.bind(this), 0);
     },
 
     prepareHeader: function prepareHeader(props, state) {
@@ -418,7 +422,7 @@ module.exports = _react2.default.createClass({
         var loadMask;
 
         if (props.loadMaskOverHeader) {
-            loadMask = _react2.default.createElement(LoadMask, { visible: props.loading });
+            loadMask = _react2.default.createElement(_reactLoadMask2.default, { visible: props.loading });
         }
 
         var paginationToolbar;
@@ -885,7 +889,7 @@ module.exports = _react2.default.createClass({
                     defaultLoading: true
                 });
 
-                var errorFn = (function (err) {
+                var errorFn = function (err) {
                     if (props.onDataSourceError) {
                         props.onDataSourceError(err);
                     }
@@ -893,13 +897,13 @@ module.exports = _react2.default.createClass({
                     this.setState({
                         defaultLoading: false
                     });
-                }).bind(this);
+                }.bind(this);
 
                 var noCatchFn = dataSource['catch'] ? null : errorFn;
 
                 dataSource = dataSource.then(function (response) {
                     return response && typeof response.json == 'function' ? response.json() : response;
-                }).then((function (json) {
+                }).then(function (json) {
 
                     if (props.onDataSourceSuccess) {
                         props.onDataSourceSuccess(json);
@@ -932,7 +936,7 @@ module.exports = _react2.default.createClass({
                     }
 
                     this.setState(newState);
-                }).bind(this), noCatchFn);
+                }.bind(this), noCatchFn);
 
                 if (dataSource['catch']) {
                     dataSource['catch'](errorFn);
@@ -1069,7 +1073,7 @@ module.exports = _react2.default.createClass({
 
     onColumnResizeDragStart: function onColumnResizeDragStart(config) {
 
-        var domNode = this.getDOMNode();
+        var domNode = (0, _reactDom.findDOMNode)(this);
         var region = Region.from(domNode);
 
         this.resizeProxyLeft = config.resizeProxyLeft - region.left;
@@ -1092,7 +1096,7 @@ module.exports = _react2.default.createClass({
 
         if (horizScrollbar && this.state.scrollLeft) {
 
-            setTimeout((function () {
+            setTimeout(function () {
                 //FF needs this, since it does not trigger scroll event when scrollbar dissapears
                 //so we might end up with grid content not visible (to the left)
 
@@ -1100,7 +1104,7 @@ module.exports = _react2.default.createClass({
                 if (domNode && !domNode.scrollLeft) {
                     this.handleScrollLeft(0);
                 }
-            }).bind(this), 1);
+            }.bind(this), 1);
         }
 
         var props = this.props;
